==> Transcriptomics/1. Quality_Control_CSF.Rmd <==
---
title: "Quality control CSF"
author: "Stijn Swinnen and Jarne Beliën"
date: "24-04-2023"
output:
  html_document:
    toc: TRUE
    toc_float: TRUE
---

*Laboratory for Neuroimmunology; University of Leuven*


**Based on (for the most part):**

- Satija Lab, [Seurat](https://satijalab.org/seurat): Vignettes.

- University of Cambridge, Wellcome Sanger Institute: [single-cell course](https://www.singlecellcourse.org/index.html).

- Harvard Chan Bioinformatics Core (HBC): [single-cell course](https://hbctraining.github.io/scRNA-seq).

- GitHub: SoupX; Detailed vignette.

- GitHub: scDblFinder; Detailed vignette.

- GitHub: harmony; Detailed vignette.

- Bioconductor: SingleR; Detailed vignette and book.


*References to specific sources (e.g., articles and GitHub community pages) are provided in this script.*


**Overview quality control (QC) parameters:**

- Cell-free (m)RNA contamination removed (SoupX).

- Doublets removed (scDblFinder).

- Number cells in which feature (gene) detected = 5.

- Minimum features (genes) per cell = 200.

- Maximum percentage mitochondrial RNA (mtRNA) per cell = 15 (scRNA-seq) or 5 (snRNA-seq).

- IGHV, IGKV, IGLV, TRAV, TRBV, TRDV and TRGV removed.

- SCTransform used to reduce technical noise.

- Number principal components (PCs) = 30.

- Percentage mtRNA and difference S-G2/M regressed out.

- Batch effect correction and integration with Harmony.

- Resolution clusters = 0.8.


**Required packages (in alphabetical order):**

```{r setup, Print Options}
knitr::opts_chunk$set(warning = FALSE, message = FALSE)
```

```{r setup, Libraries}

library(celldex)
library(clustree)
library(cowplot)
library(data.table)
library(plyr) # Needs to be loaded before dplyr
library(dplyr)
library(foreach)
library(ggplot2)
library(harmony)
library(pdfCluster)
library(purrr)
library(RColorBrewer)
library(scDblFinder)
library(Seurat)
library(SingleCellExperiment)
library(SingleR)
library(SoupX)
library(stringr)
library(tibble)
library(tidyr)

setwd("L:/GBW-0068_Neuroimmunology/ScRNA-seq/Paper CHIT1/1. Quality control/CSF")
```


Droplet-based single-cell or single-nucleus RNA-sequencing (scRNA-seq or snRNA-seq) assumes that each **droplet** contains the transcriptome of a singular cell or nucleus; issues to resolve (# 1 and # 2):

- Cell-free (ambient) (m)RNA contamination due to cell lysis.

- Empty droplets; no cell present.

- Doublets/multiplets; multiple cells present.


# 1. Removal cell-free (ambient) (m)RNA contamination due to cell lysis

*Reference SoupX: Young et al. (2020). SoupX removes ambient RNA contamination from droplet-based single-cell RNA-sequencing data. GigaScience.*

- *load10X*. Loads **unfiltered** data (raw count matrices); to profile composition background contamination in empty droplets. Identifies which droplets contain cells (based on Cell Ranger). Removes empty droplets afterwards.

- *autoEstCont*. Calculates contamination fraction at **cluster level**. Default contaminationRange = c(0.01, 0.8). Default priorRho = 0.05 and default priorRhoStdDev (priorRho standard deviation) = 0.1; -> postRho = contamination fraction (most common 0-10%).

- *adjustCounts*. Does not remove genes; subtracts background **counts** for each gene.

**Batch 1:**
```{r, fig.dim = c(10, 7)}

for (i in c("01", "02", "03", "07")) {
  Unfiltered_Data = load10X(paste0("L:/GBW-0068_Neuroimmunology/ScRNA-seq/Experiments/CSF 1 06-12-2021/2. Cell Ranger output/GEX_", i, "/SoupX/outs")) %>%
    autoEstCont() %>%
    adjustCounts()
  assign(paste0("CSF_", i), Unfiltered_Data)
}
```

**Batch 2:**
```{r, fig.dim = c(10, 7)}

for (i in c("08", "09", "11", "12", "14")) {
  Unfiltered_Data = load10X(paste0("L:/GBW-0068_Neuroimmunology/ScRNA-seq/Experiments/CSF 2 11-05-2022/2. Cell Ranger output/GEX_", i, "/SoupX/outs")) %>%
    autoEstCont() %>%
    adjustCounts()
  assign(paste0("CSF_", i), Unfiltered_Data)
}
```

**Batch 3:**
```{r, fig.dim = c(10, 7)}

for (i in c("16", "17")) {
  Unfiltered_Data = load10X(paste0("L:/GBW-0068_Neuroimmunology/ScRNA-seq/Experiments/CSF 3 23-11-2022/2. Cell Ranger output/GEX_", i, "/SoupX/outs")) %>%
    autoEstCont() %>%
    adjustCounts()
  assign(paste0("CSF_", i), Unfiltered_Data)
}
```

Output = sparse and **corrected count matrix** to be used instead of original matrix in downstream analysis. Cell count output SoupX = Cell Ranger (filtered count matrices).


# 2. Identification doublets

*Reference scDblFinder: Germain et al. (2021). Doublet identification in single-cell sequencing data using scDblFinder. F1000Research.* 

-> Benchmarked in: *Xi et al. (2021). Benchmarking computational doublet-detection methods for single-cell RNA-sequencing data (in addendum). Cell systems.*

- *scDblFinder*. In particular **heterotypic** doublets. Default clusters = NULL (random artificial doublets = homo- and heterotypic); changed to TRUE (intercluster artificial doublets = heterotypic), because homotypic doublets are unidentifiable based on their transcriptome (do not form intermediate clusters/phenotypes) and therefore erroneous call of actual singlets as doublets is possible. Default artificialDoublets = 5,000. Default dbr (doublet rate) = 1%/1,000 cells (most common 10-20%). Default dbr.sd (doublet rate standard deviation) = 40% dbr.

```{r, fig.dim = c(10, 7)}

CSF <- list(CSF_01, CSF_02, CSF_03, CSF_07, CSF_08, CSF_09, CSF_11, CSF_12, CSF_14, CSF_16, CSF_17)
CSF_Names <- list("CSF_01", "CSF_02", "CSF_03", "CSF_07", "CSF_08", "CSF_09", "CSF_11", "CSF_12", "CSF_14", "CSF_16", "CSF_17")

invisible(foreach(i = CSF, j = CSF_Names) %do% {
  Doublets <- scDblFinder(i, clusters = TRUE)
  assign(paste0("Doublets_", j), Doublets)
})

Doublets_CSF <- list(Doublets_CSF_01, Doublets_CSF_02, Doublets_CSF_03, Doublets_CSF_07, Doublets_CSF_08, Doublets_CSF_09, Doublets_CSF_11, Doublets_CSF_12, Doublets_CSF_14, Doublets_CSF_16, Doublets_CSF_17)

invisible(foreach(i = Doublets_CSF, j = CSF_Names) %do% {
  Doublets <- data.frame("Barcode" = rownames(colData(i)), "droplet.class" = i$scDblFinder.class, "doublet.score" = i$scDblFinder.score) %>%
    remove_rownames() %>%
    column_to_rownames("Barcode")
  assign(paste0("Doublets_", j), Doublets)
})
head(Doublets)
```

Output = **data frame** to add to metadata Seurat object in # 3.


# 3. Creation Seurat object

- *CreateSeuratObject*. Default assay = "RNA". Default min.cells = 0. Default min.features = 0. It is recommended to first use low values for min.cells and min.features; retain most features (genes) and cells (https://github.com/satijalab/seurat/issues/3812). In-depth quality control (QC) will ensue in # 4.

```{r, fig.dim = c(10, 7)}

invisible(foreach(i = CSF, j = CSF_Names) %do% {
  Seurat_Object <- CreateSeuratObject(counts = i, project = j, min.cells = 5, min.features = 100)
  assign(j, Seurat_Object)
})
```

Allocate droplet class ("singlet" or "doublet") and doublet score:
```{r, fig.dim = c(10, 7)}

CSF_01 <- AddMetaData(CSF_01, Doublets_CSF_01)
CSF_02 <- AddMetaData(CSF_02, Doublets_CSF_02)
CSF_03 <- AddMetaData(CSF_03, Doublets_CSF_03)
CSF_07 <- AddMetaData(CSF_07, Doublets_CSF_07)
CSF_08 <- AddMetaData(CSF_08, Doublets_CSF_08)
CSF_09 <- AddMetaData(CSF_09, Doublets_CSF_09)
CSF_11 <- AddMetaData(CSF_11, Doublets_CSF_11)
CSF_12 <- AddMetaData(CSF_12, Doublets_CSF_12)
CSF_14 <- AddMetaData(CSF_14, Doublets_CSF_14)
CSF_16 <- AddMetaData(CSF_16, Doublets_CSF_16)
CSF_17 <- AddMetaData(CSF_17, Doublets_CSF_17)
```

Allocate batch ("Batch 1", "Batch 2" or "Batch 3"):
```{r, fig.dim = c(10, 7)}

CSF_01[["batch"]] <- "Batch 1"
CSF_02[["batch"]] <- "Batch 1"
CSF_03[["batch"]] <- "Batch 1"
CSF_07[["batch"]] <- "Batch 1"
CSF_08[["batch"]] <- "Batch 2"
CSF_09[["batch"]] <- "Batch 2"
CSF_11[["batch"]] <- "Batch 2"
CSF_12[["batch"]] <- "Batch 2"
CSF_14[["batch"]] <- "Batch 2"
CSF_16[["batch"]] <- "Batch 3"
CSF_17[["batch"]] <- "Batch 3"
```

Allocate sequencing technique ("scRNA-seq" or "snRNA-seq"):
```{r, fig.dim = c(10, 7)}

CSF_01[["sequencing.technique"]] <- "scRNA-seq"
CSF_02[["sequencing.technique"]] <- "scRNA-seq"
CSF_03[["sequencing.technique"]] <- "scRNA-seq"
CSF_07[["sequencing.technique"]] <- "scRNA-seq"
CSF_08[["sequencing.technique"]] <- "scRNA-seq"
CSF_09[["sequencing.technique"]] <- "scRNA-seq"
CSF_11[["sequencing.technique"]] <- "scRNA-seq"
CSF_12[["sequencing.technique"]] <- "scRNA-seq"
CSF_14[["sequencing.technique"]] <- "scRNA-seq"
CSF_16[["sequencing.technique"]] <- "scRNA-seq"
CSF_17[["sequencing.technique"]] <- "scRNA-seq"
```

Allocate subject ("MS_0...", "MS_1...", "MS_2...", "MS_3..." or "MS_4..."):
```{r, fig.dim = c(10, 7)}

CSF_01[["subject"]] <- "MS_0_1"
CSF_02[["subject"]] <- "MS_0_2"
CSF_03[["subject"]] <- "MS_0_3"
CSF_07[["subject"]] <- "MS_0_4"
CSF_08[["subject"]] <- "MS_0_5"
CSF_09[["subject"]] <- "MS_0_6"
CSF_11[["subject"]] <- "MS_0_7"
CSF_12[["subject"]] <- "MS_0_8"
CSF_14[["subject"]] <- "MS_0_9"
CSF_16[["subject"]] <- "MS_0_10"
CSF_17[["subject"]] <- "MS_0_11"
```

Allocate disease course ("RRMS", "PPMS", "SPMS", "PMS", "MS" or "NA"):
```{r, fig.dim = c(10, 7)}

CSF_01[["disease.course"]] <- "RRMS"
CSF_02[["disease.course"]] <- "RRMS"
CSF_03[["disease.course"]] <- "RRMS"
CSF_07[["disease.course"]] <- "RRMS"
CSF_08[["disease.course"]] <- "RRMS"
CSF_09[["disease.course"]] <- "RRMS"
CSF_11[["disease.course"]] <- "RRMS"
CSF_12[["disease.course"]] <- "RRMS"
CSF_14[["disease.course"]] <- "RRMS"
CSF_16[["disease.course"]] <- "RRMS"
CSF_17[["disease.course"]] <- "RRMS"
```

Allocate tissue ("CSF", "Brain", "Spinal cord" or "NA"):
```{r, fig.dim = c(10, 7)}

CSF_01[["tissue"]] <- "CSF"
CSF_02[["tissue"]] <- "CSF"
CSF_03[["tissue"]] <- "CSF"
CSF_07[["tissue"]] <- "CSF"
CSF_08[["tissue"]] <- "CSF"
CSF_09[["tissue"]] <- "CSF"
CSF_11[["tissue"]] <- "CSF"
CSF_12[["tissue"]] <- "CSF"
CSF_14[["tissue"]] <- "CSF"
CSF_16[["tissue"]] <- "CSF"
CSF_17[["tissue"]] <- "CSF"
```

Allocate tissue subtype ("White matter", "Grey matter", "White and grey matter" or "NA"):
```{r, fig.dim = c(10, 7)}

CSF_01[["tissue.subtype"]] <- "NA"
CSF_02[["tissue.subtype"]] <- "NA"
CSF_03[["tissue.subtype"]] <- "NA"
CSF_07[["tissue.subtype"]] <- "NA"
CSF_08[["tissue.subtype"]] <- "NA"
CSF_09[["tissue.subtype"]] <- "NA"
CSF_11[["tissue.subtype"]] <- "NA"
CSF_12[["tissue.subtype"]] <- "NA"
CSF_14[["tissue.subtype"]] <- "NA"
CSF_16[["tissue.subtype"]] <- "NA"
CSF_17[["tissue.subtype"]] <- "NA"
```

Allocate pathology ("Lesion", "Lesion core", "Lesion edge", "Periplaque", "Normal-appearing matter" or "NA"):
```{r, fig.dim = c(10, 7)}

CSF_01[["pathology"]] <- "NA"
CSF_02[["pathology"]] <- "NA"
CSF_03[["pathology"]] <- "NA"
CSF_07[["pathology"]] <- "NA"
CSF_08[["pathology"]] <- "NA"
CSF_09[["pathology"]] <- "NA"
CSF_11[["pathology"]] <- "NA"
CSF_12[["pathology"]] <- "NA"
CSF_14[["pathology"]] <- "NA"
CSF_16[["pathology"]] <- "NA"
CSF_17[["pathology"]] <- "NA"
```

Allocate pathology activity ("Acute active", "Acute inactive", "Chronic active", "Chronic inactive", "Acute/chronic active", "Acute/chronic inactive" or "NA"):
```{r, fig.dim = c(10, 7)}

CSF_01[["pathology.activity"]] <- "NA"
CSF_02[["pathology.activity"]] <- "NA"
CSF_03[["pathology.activity"]] <- "NA"
CSF_07[["pathology.activity"]] <- "NA"
CSF_08[["pathology.activity"]] <- "NA"
CSF_09[["pathology.activity"]] <- "NA"
CSF_11[["pathology.activity"]] <- "NA"
CSF_12[["pathology.activity"]] <- "NA"
CSF_14[["pathology.activity"]] <- "NA"
CSF_16[["pathology.activity"]] <- "NA"
CSF_17[["pathology.activity"]] <- "NA"

CSF <- list(CSF_01, CSF_02, CSF_03, CSF_07, CSF_08, CSF_09, CSF_11, CSF_12, CSF_14, CSF_16, CSF_17)

for (i in CSF) {
  summary(i@meta.data) %>%
    print()
}
```

Output = **Seurat objects** with assay **RNA** and slot **counts** (raw). Next, we **merge** these separate Seurat objects into a singular Seurat object. Because the same bar code can be used in different samples, we add a sample-specific prefix to each bar code.

```{r, fig.dim = c(10, 7)}

Merged_CSF <- merge(CSF_01, c(CSF_02, CSF_03, CSF_07, CSF_08, CSF_09, CSF_11, CSF_12, CSF_14, CSF_16, CSF_17 ), add.cell.ids = CSF_Names)

head(Merged_CSF)
tail(Merged_CSF)
summary(Merged_CSF@meta.data)
DefaultAssay(Merged_CSF)
```


# 4. Determination quality control (QC) parameters to identify high-quality cells


Let us now define **High-quality** cells:

- High number features (genes)/cell.

- Low percentage mitochondrial RNA (mtRNA)/cell.

- Droplet class = singlets.

First, let's calculate the percentage of mitochondrial RNA (mtRNA) and ribosomal (rRNA):
```{r, fig.dim = c(10, 7)}

Merged_CSF[["percent.mt"]] <- PercentageFeatureSet(Merged_CSF, pattern = "^MT-")
Merged_CSF[["percent.rb"]] <- PercentageFeatureSet(Merged_CSF, pattern = "^RP[SL]")

summary(Merged_CSF@meta.data)
```

Next we plot these features:

FeatureScatter:
```{r, fig.dim = c(10, 7)}

for (i in CSF_Names) {
  Subset <- subset(Merged_CSF, orig.ident == i)
  print(i)
  summary(Subset@meta.data) %>%
    print()
}

VlnPlot(Merged_CSF, features = "nCount_RNA")
VlnPlot(Merged_CSF, features = "nFeature_RNA")
VlnPlot(Merged_CSF, features = "percent.mt")
VlnPlot(Merged_CSF, features = "percent.rb")
VlnPlot(Merged_CSF, features = "doublet.score")
```

Correlation QC parameters:
```{r, fig.dim = c(10, 7)}

Plot1 <- FeatureScatter(Merged_CSF, feature1 = "nCount_RNA", feature2 = "nFeature_RNA")
Plot2 <- FeatureScatter(Merged_CSF, feature1 = "nFeature_RNA", feature2 = "percent.mt")
Plot3 <- FeatureScatter(Merged_CSF, feature1 = "nFeature_RNA", feature2 = "percent.rb")
Plot4 <- FeatureScatter(Merged_CSF, feature1 = "percent.mt", feature2 = "percent.rb")
Plot5 <- FeatureScatter(Merged_CSF, feature1 = "nFeature_RNA", feature2 = "doublet.score")

Plot1
Plot2
Plot3
Plot4
Plot5
```

Now, we add a QC column (as a subset; no low-quality cells removed yet) to the metadata of the merged Seurat object and specify our **pass or fail criteria**. We will implement these QC criteria in # 8.

*Reference percentage mtRNA: Osorio et al. (2021). Systematic determination of the mitochondrial proportion in human and mice tissues for single-cell RNA-sequencing data quality control. Bioinformatics.* -> **< 15%**, based on fig. 1c; monocyte-derived macrophages used as estimator for microglia.

```{r, fig.dim = c(10, 7)}

Merged_CSF[["quality.control"]] <- ifelse(Merged_CSF$nFeature_RNA > 200 & Merged_CSF$percent.mt < 15 & Merged_CSF$droplet.class == "singlet", "Pass", "Fail")
```

Overview "Fail" and "Pass" per sample:
```{r, fig.dim = c(10, 7)}

Table1 <- table(Merged_CSF$orig.ident, Merged_CSF$quality.control)
Table2 <- colSums(Table1)
Table3 <- prop.table(Table2)
Table4 <- sum(Table3)
Table5 <- prop.table(Table1, margin = 1)
Table6 <- rowSums(Table5)

Table1 # "Fail" and "Pass" per sample
Table2 # Sum total "Fail" and "Pass"
Table3 # Proportion total "Fail" and "Pass"
Table4 # Sum proportions total "Fail" and "Pass"; must be 1
Table5 # Proportion "Fail" and "Pass" per sample
Table6 # Sum proportions "Fail" and "Pass" per sample; must be 1

summary(subset(Merged_CSF@meta.data, quality.control == "Pass"))

for (i in CSF_Names) {
  Subset <- subset(Merged_CSF, quality.control == "Pass" & orig.ident == i)
  print(i)
  summary(Subset@meta.data) %>%
    print()
}

VlnPlot(subset(Merged_CSF, quality.control == "Pass"), features = "nFeature_RNA")
VlnPlot(subset(Merged_CSF, quality.control == "Pass"), features = "percent.mt")
VlnPlot(subset(Merged_CSF, quality.control == "Pass"), features = "percent.rb")
VlnPlot(subset(Merged_CSF, quality.control == "Pass"), features = "doublet.score")
```


Sources (unwanted) **variation** (# 5, # 6 and # 7):

- Variable fragments B and T cell receptors (BCR and TCR); induced subject-specific clusters in our hands.

- Differences cell cycle phase.

- Batch effects.


# 5. Removal variable fragments B and T cell receptors (BCR and TCR)

```{r, fig.dim = c(10, 7)}

Counts <- GetAssayData(Merged_CSF, assay = "RNA", slot = "counts")
Genes <- rownames(Counts)
Variable_Genes <- Genes[str_detect(Genes, pattern = "^IGHV|^IGKV|^IGLV|^TRAV|^TRBV|^TRDV|^TRGV")]
Counts <- Counts[-(which(rownames(Counts) %in% Variable_Genes)), ]
Merged_CSF <- subset(Merged_CSF, features = rownames(Counts))

Variable_Genes <- as.data.frame(Variable_Genes)
dplyr::count(Variable_Genes) # Number genes removed
Genes <- as.data.frame(Genes)
dplyr::count(Variable_Genes)/dplyr::count(Genes) # Proportion genes removed
```

Influence on number cells "Fail" or "Pass" QC criteria?
```{r, fig.dim = c(10, 7)}

Merged_CSF[["quality.control"]] <- ifelse(Merged_CSF$nFeature_RNA > 200 & Merged_CSF$percent.mt < 15 & Merged_CSF$droplet.class == "singlet", "Pass", "Fail")
Table1 <- table(Merged_CSF$orig.ident, Merged_CSF$quality.control)
Table1
```


# 6. Exploration cell cycle phase variation

It is recommended to **normalize** the data before assessment of the cell cycle score (https://github.com/satijalab/seurat/issues/1679). The cell cycle score is assigned to each **individual cell** - independent of other cells -, which means this function can be used before implementation of the QC criteria in # 8 (i.e., no influence of cells that will be removed).

- *NormalizeData*. Corrects for sequencing depth; "Merged_CSF@assays$RNA@data". Default normalization.method = "LogNormalize". Default scale.factor = 10,000 (as if all cells contain 10,000 transcripts).

- *FindVariableFeatures*. For principal component analysis (PCA). Default selection.method = "vst". Default nfeatures = 2,000.

- *ScaleData*. Corrects for differences in variation gene expression, Z-score; "Merged_CSF@assays$RNA@scale.data". Mean expression = 0; expression variance across cells = 1 (i.e., equal variation gene expression). Default features = NULL (all genes).

```{r, fig.dim = c(10, 7)}

Merged_CSF <- NormalizeData(Merged_CSF, verbose = FALSE)
Merged_CSF <- FindVariableFeatures(Merged_CSF, verbose = FALSE)
Top10 <- head(VariableFeatures(Merged_CSF), 10)
Top10 # Unless removed, dominated by variable fragments BCR and TCR
Plot <- VariableFeaturePlot(Merged_CSF)
LabelPoints(plot = Plot, points = Top10, repel = TRUE, xnudge = 0, ynudge = 0)
Merged_CSF <- ScaleData(Merged_CSF, features = rownames(Merged_CSF), verbose = FALSE)
```

Output = **Seurat object** with assay **RNA** and slots **counts** (raw), **data** (log-normalized) and **scale.data** (scaled).

Next, we evaluate cell cycle scores in our cells:
```{r, fig.dim = c(10, 7)}

cc.genes.updated.2019
s.genes <- cc.genes.updated.2019$s.genes
g2m.genes <- cc.genes.updated.2019$g2m.genes
Merged_CSF <- CellCycleScoring(Merged_CSF, s.features = s.genes, g2m.features = g2m.genes, set.ident = TRUE) # Active.ident changed from samples (orig.ident) to cell cycle phases

summary(Merged_CSF@meta.data)
```

Cell cycle statistics:
```{r, fig.dim = c(10, 7)}

Table1 <- table(Merged_CSF$orig.ident, Merged_CSF$Phase)
Table2 <- prop.table(Table1, margin = 1)
Table3 <- rowSums(Table2)

Table1 # Cell cycle phases per sample
Table2 # Proportion cell cycle phases per sample
Table3 # Sum proportions cell cycle phases per sample; must be 1
```

Since we only use cell cycle genes as input for the principal components (PCs), there is an insufficient number of genes to generate 50 PCs (default); changed to 35 PCs (https://github.com/satijalab/seurat/discussions/4567):
```{r, fig.dim = c(10, 7)}

Merged_CSF <- RunPCA(Merged_CSF, features = c(s.genes, g2m.genes), npcs = 35)
Merged_CSF@reductions
DefaultDimReduc(Merged_CSF)
PCAPlot(Merged_CSF) # Displays variation between cells due to differences in cell cycle phase
Merged_CSF <- RunPCA(Merged_CSF)
Merged_CSF@reductions
DefaultDimReduc(Merged_CSF)
PCAPlot(Merged_CSF, split.by = "Phase") # Displays influence cell cycle genes on PC_1 and PC_2
ElbowPlot(Merged_CSF)
Merged_CSF <- FindNeighbors(Merged_CSF)
Merged_CSF <- FindClusters(Merged_CSF) # Active.ident changed from cell cycle phases to clusters
Merged_CSF <- RunUMAP(Merged_CSF, dims = 1:15, verbose = FALSE)
Merged_CSF@reductions
DefaultDimReduc(Merged_CSF)
UMAPPlot(Merged_CSF, label = TRUE)
FeaturePlot(Merged_CSF, features = "S.Score", min.cutoff = "q10", label = TRUE)
VlnPlot(Merged_CSF, features = "S.Score")
FeaturePlot(Merged_CSF, features = "G2M.Score", min.cutoff = "q10", label = TRUE)
VlnPlot(Merged_CSF, features = "G2M.Score")
```

There might be some biologically relevant variation between **cycling** - S and G2M genes - and **non-cycling cells** - G1 genes - (e.g. plasmablasts). Thus, we want to keep this source of variation in our data. The difference between cells with expression of S genes or G2M genes, however, is minimal (and probably an unwanted source of variation). Therefore, we regress this difference out in # 9.

```{r, fig.dim = c(10, 7)}

Merged_CSF[["cc.difference"]] <- Merged_CSF$S.Score - Merged_CSF$G2M.Score
summary(Merged_CSF@meta.data)
```


# 7. Estimation batch effects

Visualization of the batch effects:
```{r, fig.dim = c(10, 7)}

UMAPPlot(Merged_CSF, group.by = "batch", label = TRUE, repel = TRUE)
```


# 8. Implementation QC criteria

Which cells will be removed (some cells might be removed due to more than one criterion) under our QC criteria?

- **nFeature_RNA < 200**
```{r, fig.dim = c(10, 7)}

Fail_nFeature_RNA <- WhichCells(Merged_CSF, expression = nFeature_RNA < 200)
dplyr::count(as.data.frame(Fail_nFeature_RNA))
Subset <- subset(Merged_CSF, cells = Fail_nFeature_RNA)
Table1 <- table(Subset$old.ident, Subset$quality.control)
Table2 <- Table1/as.vector(table(Merged_CSF$old.ident))
Table3 <- table(Subset$RNA_snn_res.0.8, Subset$quality.control)
Table4 <- Table3/as.vector(table(Merged_CSF$RNA_snn_res.0.8))

Table1 # "Fail" nFeature_RNA per sample
Table2 # Proportion "Fail" nFeature_RNA per sample
Table3 # "Fail" nFeature_RNA per cluster
Table4 # Proportion "Fail" nFeature_RNA per cluster

UMAPPlot(Merged_CSF, split.by = "quality.control", label = TRUE, cells.highlight = Fail_nFeature_RNA) + NoLegend()
```

- **percent.mt > 15**
```{r, fig.dim = c(10, 7)}

Fail_percent.mt <- WhichCells(Merged_CSF, expression = percent.mt > 15)
dplyr::count(as.data.frame(Fail_percent.mt))
Subset <- subset(Merged_CSF, cells = Fail_percent.mt)
Table1 <- table(Subset$old.ident, Subset$quality.control)
Table2 <- Table1/as.vector(table(Merged_CSF$old.ident))
Table3 <- table(Subset$RNA_snn_res.0.8, Subset$quality.control)
Table4 <- Table3/as.vector(table(Merged_CSF$RNA_snn_res.0.8))

Table1 # "Fail" percent.mt per sample
Table2 # Proportion "Fail" percent.mt per sample
Table3 # "Fail" percent.mt per cluster
Table4 # Proportion "Fail" percent.mt per cluster

UMAPPlot(Merged_CSF, split.by = "quality.control", label = TRUE, cells.highlight = Fail_percent.mt) + NoLegend()
```

- **droplet.class == "doublet"**
```{r, fig.dim = c(10, 7)}

Fail_droplet.class <- WhichCells(Merged_CSF, expression = droplet.class == "doublet")
dplyr::count(as.data.frame(Fail_droplet.class))
Subset <- subset(Merged_CSF, cells = Fail_droplet.class)
Table1 <- table(Subset$old.ident, Subset$quality.control)
Table2 <- Table1/as.vector(table(Merged_CSF$old.ident))
Table3 <- table(Subset$RNA_snn_res.0.8, Subset$quality.control)
Table4 <- Table3/as.vector(table(Merged_CSF$RNA_snn_res.0.8))

Table1 # "Fail" droplet.class per sample
Table2 # Proportion "Fail" droplet.class per sample
Table3 # "Fail" droplet.class per cluster
Table4 # Proportion "Fail" droplet.class per cluster

UMAPPlot(Merged_CSF, split.by = "quality.control", label = TRUE, cells.highlight = Fail_droplet.class) + NoLegend()
```

Overview "Fail" and "Pass" per cluster (cluster information was not yet available in # 4):
```{r, fig.dim = c(10, 7)}

Table1 <- table(Merged_CSF$RNA_snn_res.0.8, Merged_CSF$quality.control)
Table2 <- colSums(Table1)
Table3 <- prop.table(Table2)
Table4 <- sum(Table3)
Table5 <- prop.table(Table1, margin = 1)
Table6 <- rowSums(Table5)

Table1 # "Fail" and "Pass" per cluster
Table2 # Sum total "Fail" and "Pass"
Table3 # Proportion total "Fail" and "Pass"
Table4 # Sum proportions total "Fail" and "Pass"; must be 1
Table5 # Proportion "Fail" and "Pass" per cluster
Table6 # Sum proportions "Fail" and "Pass" per cluster; must be 1
```

Implementation QC criteria:
```{r, fig.dim = c(10, 7)}

Merged_CSF <- subset(Merged_CSF, quality.control == "Pass")

summary(Merged_CSF@meta.data)

for (i in CSF_Names) {
  Subset <- subset(Merged_CSF, orig.ident == i)
  print(i)
  summary(Subset@meta.data) %>%
    print()
}
```


# 9. Data normalization and variance stabilization with *SCTransform*

*Reference SCTransform: Hafemeister and Satija (2019). Normalization and variance stabilization of single-cell RNA-seq data using regularized negative binomial regression. Genome Biology.*

*SCTransform* is beneficial to find **"rare" (low-frequency) cell populations**. In comparison to *NormalizeData*, *SCTransform* is less susceptible to technical noise and therefore allows more variable features (genes; 2,000 -> 3,000) and more PCs. In addition, we **regress out** (not remove) the percentage of mtRNA (common practice) and the difference between the expression of S and G2M genes (difference between cycling and non-cycling cells remains).

- *SCTransform*. *NormalizeData*, *FindVariableFeatures* and *ScaleData* in one optimized function; "Split_CSF@assays$SCT". Default assay = "RNA". Default residual.features = NULL (all genes). Default variable.features.n = 3,000.

We split the merged Seurat object again before *SCTransfrom*: (https://github.com/satijalab/seurat/issues/5306)
```{r, fig.dim = c(10, 7)}

Split_CSF <- SplitObject(Merged_CSF, split.by = "orig.ident")
for (i in 1:length(Split_CSF)) {
  Split_CSF[[i]] <- SCTransform(Split_CSF[[i]], vars.to.regress = c("percent.mt", "cc.difference"), verbose = FALSE)
}
```

Check assays stored in Seurat objects (e.g., CSF_01):
```{r, fig.dim = c(10, 7)}

Split_CSF$CSF_01@assays
DefaultAssay(Split_CSF$CSF_01)
```

Output = **Seurat objects** with additional assay **SCT** and slots **counts** (corrected; reverse-transformed from Pearson residuals in scale.data and sequence-depth corrected), **data** (corrected; log-normalized counts) and **scale.data** (Pearson residuals).


# 10. Data integration and batch effect correction with Harmony

*Reference Harmony: Korsunsky et al (2019). Fast, sensitive and accurate integration of single-cell data with Harmony. Nature Methods.*

We specify that we used **SCTransform** with 3,000 variable features (genes).

- *SelectIntegrationFeatures*. Default nfeatures = 2,000.

- *RunPCA*. Default features = NULL (variable features). Default npcs (number of PCs) = 50.

Harmony integration transforms principal components (PCs) and seeks clusters with maximum "diversity" based on k-means, i.e. whether each cluster contains balanced amounts of cells from each batch (e.g., donor, condition, tissue,...). Harmony tries to correct the imbalance or **batch effects**. To do so, a merged and normalized Seurat object with defined variable features and PCs is required (https://hbctraining.github.io/scRNA-seq_online/lessons/06a_integration_harmony.html).

-> Benchmarked in *Tran et al. (2020). A benchmark of batch-effect correction methods for single-cell RNA-sequencing data. Genome Biology.* and *Luecken et al. (2022). Benchmarking atlas-level data integration in single-cell genomics. Nature Methods.* 

We treat each **sample** as a different batch in order to integrate on the level of the individual samples.
```{r, fig.dim = c(10, 7)}

Integration_Features <- SelectIntegrationFeatures(Split_CSF, nfeatures = 3000)
Merged_CSF <- merge(Split_CSF[[1]], Split_CSF[2:length(Split_CSF)])
VariableFeatures(Merged_CSF) <- Integration_Features
Merged_CSF <- RunPCA(Merged_CSF, assay = "SCT")

Harmonized_CSF <- RunHarmony(Merged_CSF, group.by.vars = "orig.ident", reduction = "pca", assay.use = "SCT", reduction.save = "harmony")
Harmonized_CSF@reductions
DefaultDimReduc(Harmonized_CSF)
DefaultAssay(Harmonized_CSF)
```

Output = batch-corrected **Seurat object** with additional reduction **harmony** (in addition to **pca**).


# 11. Clustering

Selection of the number of "Harmony" PCs:

- *ElbowPlot*. Default reduction = "pca".
```{r, fig.dim = c(10, 7)}

ElbowPlot(Harmonized_CSF, ndims = 50, reduction = "harmony")
```

*SCTransform* is less susceptible to technical noise than *NormalizeData*, *FindVariableFeatures* and *ScaleData*. Therefore, we can choose a higher number of "Harmony" PCs without incorporating - too much - technical sources of variation. We decide to use the first **30** "Harmony" PCs.

- *FindNeighbors*. Default reduction = "pca".

- *FindClusters*. Default resolution = 0.8.

- *RunUMAP*. Default reduction = "pca".

```{r, fig.dim = c(10, 7)}

Harmonized_CSF <- FindNeighbors(Harmonized_CSF, reduction = "harmony", dims = 1:30)
Harmonized_CSF <- FindClusters(Harmonized_CSF)
Harmonized_CSF@commands$FindClusters # Parameters used to determine clusters
Harmonized_CSF <- RunUMAP(Harmonized_CSF, reduction = "harmony", dims = 1:30, verbose = FALSE)
Harmonized_CSF@reductions
DefaultDimReduc(Harmonized_CSF)
UMAPPlot(Harmonized_CSF, label = TRUE) # Resolution 0.8
```

Visualization batch effects post-integration:
```{r, fig.dim = c(10, 7)}

UMAPPlot(Harmonized_CSF, group.by = "batch", label = TRUE, repel = TRUE)
```


# 12. Preliminary broad cluster annotation

We probe whether we find the **expected (large) cell populations**. With regard to differential gene expression (DGE) and visualization of differentially expressed genes (DEGs), it is recommended to use the **RNA assay** - not the SCT assay (https://github.com/satijalab/seurat/issues/4081; https://github.com/satijalab/seurat/issues/4082).


## 12.1 Manual annotation

```{r, fig.dim = c(10, 7)}

DefaultAssay(Harmonized_CSF)
DefaultAssay(Harmonized_CSF) <- "RNA"
DefaultAssay(Harmonized_CSF)
```

**Myeloid cells**

- General myeloid cells: Upregulation PTPRC, CSF1R, P2RY12, ITGAM and LYZ.

```{r, fig.dim = c(10, 7)}

FeaturePlot(Harmonized_CSF, features = c("PTPRC", "CSF1R", "P2RY12", "ITGAM", "LYZ"), label = TRUE)
VlnPlot(Harmonized_CSF, features = c("PTPRC", "CSF1R", "P2RY12", "ITGAM", "LYZ"))
```

**Dendritic cells**

- Plasmacytoid dendritic cells (pDC): Upregulation TCF4 and TNFRSF21.

- Myeloid dendritic cells 1 (mDC 1): Upregulation WDFY4, XCR1, BATF3, AXL and SIGLEC6.

- Myeloid dendritic cells 2 (mDC 2): Upregulation FCER1A, CD1C, CLEC10A and CLEC9A.

```{r, fig.dim = c(10, 7)}

FeaturePlot(Harmonized_CSF, features = c("TCF4", "TNFRSF21"), label = TRUE)
FeaturePlot(Harmonized_CSF, features = c("WDFY4", "XCR1", "BATF3", "AXL", "SIGLEC6"), label = TRUE)
FeaturePlot(Harmonized_CSF, features = c("FCER1A", "CD1C", "CLEC10A", "CLEC9A"), label = TRUE)
```

**Granulocytes**

- Granulocytes: Upregulation S100A8 and S100A9.

```{r, fig.dim = c(10, 7)}

FeaturePlot(Harmonized_CSF, features = c("S100A8", "S100A9"), label = TRUE)
```

**Monocytes and macrophages/microglia**

- General monocytes: Upregulation VCAN, CCR2, HLA-DBP1 and CD74.

- CD14+ monocytes: Upregulation CD14.

- CD16+ monocytes: Upregulation FCGR3A (CD16).

- Macrophages/microglia: Upregulation APOE and TREM2.

- Macrophages: Upregulation EMP3 and LYVE1.

- Microglia: Upregulation CD81, P2RY12, TMEM119, SPI1 and CX3CR1.

```{r, fig.dim = c(10, 7)}

FeaturePlot(Harmonized_CSF, features = c("VCAN", "CCR2", "HLA-DBP1", "CD74"), label = TRUE)
FeaturePlot(Harmonized_CSF, features = "CD14", label = TRUE)
FeaturePlot(Harmonized_CSF, features = "FCGR3A", label = TRUE)
FeaturePlot(Harmonized_CSF, features = c("APOE", "TREM2"), label = TRUE)
VlnPlot(Harmonized_CSF, features = c("APOE", "TREM2"))
FeaturePlot(Harmonized_CSF, features = c("EMP3", "LYVE1"), label = TRUE)
FeaturePlot(Harmonized_CSF, features = c("CD81", "P2RY12", "TMEM119", "SPI1", "CX3CR1"), label = TRUE)
```


## 12.2 Automated cluster annotation with *SingleR*

*Reference SingleR: Aran et al. (2019). Reference-based analysis of lung single-cell sequencing reveals a transitional profibrotic macrophage. Nature immunology.*

We compare our manual cluster annotation (in # 12.1) with an automated cluster annotation method *SingleR* - based on Spearman correlations between our dataset and a reference dataset. We use a **reference dataset** from *celldex* (Novershtern (microarray)).

Immune reference dataset (celldex):
```{r, fig.dim = c(10, 7)}

Novershtern.ref <- celldex::NovershternHematopoieticData()
```

Convert Seurat object to SingleCellExperiment (Bioconductor):
```{r, fig.dim = c(10, 7)}

SCE <- as.SingleCellExperiment(DietSeurat(Harmonized_CSF))
SCE
```

Extract cluster information:
```{r, fig.dim = c(10, 7)}

Clusters <- SCE$seurat_clusters
```

Cluster annotation with fine labels:
```{r, fig.dim = c(10, 7)}

Novershtern.fine <- SingleR(test = SCE, ref = Novershtern.ref, labels = Novershtern.ref$label.fine, clusters = Clusters)
table(Novershtern.fine$labels) # Number clusters with same annotation; grouped
table(Novershtern.fine$pruned.labels) # Low-quality annotations removed
plotScoreHeatmap(Novershtern.fine, show.pruned = TRUE, show_colnames = TRUE)
Harmonized_CSF[["Novershtern.fine"]] <- Novershtern.fine$labels[match(Harmonized_CSF[[]]["seurat_clusters"]$seurat_clusters, rownames(Novershtern.fine))]
Harmonized_CSF <- SetIdent(Harmonized_CSF, value = "Novershtern.fine")
UMAPPlot(Harmonized_CSF, label = TRUE) + NoLegend()
```

Cell annotation with fine labels:
```{r, fig.dim = c(10, 7)}
Novershtern.fine <- SingleR(test = SCE, ref = Novershtern.ref, labels = Novershtern.ref$label.fine)
table(Novershtern.fine$labels) # Number cells with same annotation; grouped
table(Novershtern.fine$pruned.labels) # Low-quality annotations removed
Harmonized_CSF[["Novershtern.fine"]] <- Novershtern.fine$labels
Harmonized_CSF <- SetIdent(Harmonized_CSF, value = "Novershtern.fine")
UMAPPlot(Harmonized_CSF, label = TRUE, repel = TRUE) + NoLegend()
```


# 13. Preliminary differential gene expression

Expected gene expression? Aberrant gene expression? Can we identify the myeloid cluster(s) based on DEGs?

- *FindAllMarkers*. Default logfc.threshold = 0.25. Default test.use = "wilcox". Default min.pct (only test genes detected in a minimum fraction of cells) = 0.1.

```{r, fig.dim = c(10, 7)}

Harmonized_CSF <- SetIdent(Harmonized_CSF, value = Harmonized_CSF$seurat_clusters)
All_Markers <- FindAllMarkers(Harmonized_CSF, min.pct = 0.25, verbose = FALSE, only.pos = TRUE)
```

Top 10 cluster markers; sorted by average log2 fold change:
```{r, fig.dim = c(10, 7)}
options(max.print = 2000) # To overrule default max.print
as.data.frame(All_Markers %>% group_by(cluster) %>% slice_max(avg_log2FC, n = 10))
```


# 14. Finalization

```{r, fig.dim = c(10, 7)}

summary(Harmonized_CSF@meta.data)

for (i in CSF_Names) {
  Subset <- subset(Harmonized_CSF, orig.ident == i)
  print(i)
  summary(Subset@meta.data) %>%
    print()
}

UMAPPlot(Harmonized_CSF, label = TRUE)

table1 <- table(Harmonized_CSF@active.ident)
table1

str(Harmonized_CSF)
saveRDS(Harmonized_CSF, "CSF.rds")

sessionInfo()
```
